requirements.txt
pandas>=1.5.0
scikit-learn>=1.2.0
numpy>=1.24.0
scipy>=1.10.0
joblib>=1.2.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
pymorphy2>=0.9.1
matplotlib>=3.6.0
seaborn>=0.12.0
